requirements.txt
joblib==1.4.2
matplotlib==3.9.2
numpy==1.26.4
pandas==2.2.3
plotly==5.24.1
PyYAML==6.0.2
scikit-learn==1.5.2
scipy==1.13.1
tqdm==4.66.5
requests==2.32.3
